+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add additional button stories
+
 ## 0.1.7
 
  - Add OctantisInteractable handled with Components.Octantis